pyproject.toml
[build-system]
requires = ["wheel", "hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "annotation-correction-3D-boxes"
version = "0.1.0"
description = "Official code of Correcting and Quantifying Systematic Errors in 3D Box Annotations for Autonomous Driving, presented at The IEEE/CVF Winter Conference on Applications of Computer Vision 2026, by Justo Miro, A. et al."
authors = [
    {name = "Alexandre Justo Miro", email = "alexandre.justo.miro@mdu.se"}
]
license = "MIT"
readme = "README.md"
# For most dependencies, we fix the major version X and give a lower bound for the minor version Y by writing ~=X.Y
requires-python = "~=3.9"
dependencies = [
    "matplotlib~=3.9",  # 3.9.4 is the last one compatible with Python 3.9
    "numpy",  # numpy v2 not fixed due to that some packages (out of our package) still require v1
    "pandas~=2.3",
    "pillow>=11.3,<13",  # 11.3.0 is the last one compatible with Python 3.9
    "pyarrow~=21.0",
    "pymoo~=0.6",
    "pytransform3d~=3.14",
    "rerun-sdk==0.23.1",  # Last rerun-sdk version that uses numpy v1. Also, 0.24 and newer versions break our code
    "tqdm~=4.67",
    "truckscenes-devkit~=1.0"
]
